+++ modules/graphics/tests/bug_github_issue_#1080.m
@@ -1,0 +1,35 @@
+%=============================================================================
+% Copyright (c) 2017 Allan CORNET (Nelson)
+%=============================================================================
+% This file is part of the Nelson.
+%=============================================================================
+% LICENCE_BLOCK_BEGIN
+% SPDX-License-Identifier: LGPL-3.0-or-later
+% LICENCE_BLOCK_END
+%=============================================================================
+% <-- Issue URL -->
+% https://github.com/nelson-lang/nelson/issues/1080
+% <-- Short Description -->
+% LineStyle, LineWidth properties were not implemented for surface object.
+%=============================================================================
+% <--ADV-CLI MODE-->
+%=============================================================================
+f = figure();
+[X, Y, Z] = peaks(35);
+C(:, :, 1) = zeros(35);
+C(:, :, 2) = ones(35) .* linspace(0.5, 0.6, 35);
+C(:, :, 3) = ones(35) .* linspace(0, 1, 35);
+S = surf(X, Y, Z, C);
+view(2);
+S.LineStyle = 'none';
+%=============================================================================
+f = figure();
+[X, Y, Z] = peaks(35);
+C(:, :, 1) = zeros(35);
+C(:, :, 2) = ones(35) .* linspace(0.5, 0.6, 35);
+C(:, :, 3) = ones(35) .* linspace(0, 1, 35);
+S = surf(X, Y, Z, C);
+view(2);
+S.LineStyle = ':';
+S.LineWidth = 2;
+%=============================================================================